chore: minor improvements
- use fn instead of setting a variable directly
- move helper class to Colors/Surface
