+++ packages/HorizonUI/Sources/HorizonUI/Sources/Foundation/Colors/HorizonUI.Colors.Surface.swift
@@ -26,7 +26,7 @@ public extension HorizonUI.Colors {
         public let cardSecondary = Color(hexString: "#FFFDFA")
         public let divider = Color(hexString: "#E8EAEC")
         public let error = Color(hexString: "#C71F23")
-        public var institution: Color { Color(uiColor: HorizonUI.institution.color) }
+        public var institution: Color { Color(uiColor: Self.institutionWrapper.color) }
         public let inversePrimary = Color(hexString: "#273540")
         public let inverseSecondary = Color(hexString: "#0A1B2A")
         public let overlayGrey = Color(hexString: "#586874")
@@ -36,5 +36,23 @@ public extension HorizonUI.Colors {
         public let pageTertiary = Color(hexString: "#E8EAEC")
         public let warning = Color(hexString: "#CF4A00")
         public let success = Color(hexString: "#03893D")
+        
+        static let institutionWrapper = InstitutionWrapper()
+        
+        /// Helper data struct to set and get the primary brand color that is handled as Institution Color.
+        final class InstitutionWrapper: @unchecked Sendable {
+            // MARK: - Private Properties
+
+            private var _color: UIColor = .black
+            private let queue = DispatchQueue(label: "com.horizonUI.colors", attributes: .concurrent)
+
+            public var color: UIColor {
+                get {
+                    return queue.sync { _color }
+                } set {
+                    queue.async(flags: .barrier) { self._color = newValue }
+                }
+            }
+        }
     }
 }
